# backend/swagger.yaml
openapi: 3.0.0
info:
  title: Node.js Sign-up and Verification API
  description: Node.js and MySQL - API with email sign-up, verification, authentication and forgot password
  version: 1.0.0
servers:
  - url: http://localhost:4000
    description: Local development server
paths:
...
  /departments:
    get:
      summary: Get all departments
      description: Returns a list of all departments
      operationId: getAllDepartments
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of departments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Department"
              example:
                - id: 1
                  name: "IT Department"
                  description: "Handles IT operations"
                  code: "IT"
                  location: "Floor 3"
                  accountId: 1
                  created: "2024-06-01T12:00:00Z"
                  updated: "2024-06-01T12:00:00Z"
                  isActive: true
                - id: 2
                  name: "HR Department"
                  description: "Handles HR operations"
                  code: "HR"
                  location: "Floor 2"
                  accountId: 2
                  created: "2024-06-01T12:00:00Z"
                  updated: "2024-06-01T12:00:00Z"
                  isActive: true
        "401":
          $ref: "#/components/responses/UnauthorizedError"
    post:
      summary: Create a new department
      description: Admin only - Create a new department
      operationId: createDepartment
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDepartmentRequest"
            example:
              name: "IT Department"
              description: "Handles IT operations"
              code: "IT"
              location: "Floor 3"
              accountId: 1
      responses:
        "200":
          description: Department created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Department"
              example:
                id: 1
                name: "IT Department"
                description: "Handles IT operations"
                code: "IT"
                location: "Floor 3"
                accountId: 1
                created: "2024-06-01T12:00:00Z"
                updated: "2024-06-01T12:00:00Z"
                isActive: true
        "400":
          description: Invalid input
        "401":
          $ref: "#/components/responses/UnauthorizedError"
  /departments/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
        description: Department ID
    get:
      summary: Get department by ID
      description: Returns a single department by ID
      operationId: getDepartmentById
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Department details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Department"
              example:
                id: 1
                name: "IT Department"
                description: "Handles IT operations"
                code: "IT"
                location: "Floor 3"
                accountId: 1
                created: "2024-06-01T12:00:00Z"
                updated: "2024-06-01T12:00:00Z"
                isActive: true
        "404":
          $ref: "#/components/responses/NotFoundError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
    put:
      summary: Update department
      description: Admin only - Update department details
      operationId: updateDepartment
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDepartmentRequest"
            example:
              name: "IT Department Updated"
              description: "Handles IT operations and support"
              code: "IT"
              location: "Floor 4"
              accountId: 1
      responses:
        "200":
          description: Department updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Department"
              example:
                id: 1
                name: "IT Department Updated"
                description: "Handles IT operations and support"
                code: "IT"
                location: "Floor 4"
                accountId: 1
                created: "2024-06-01T12:00:00Z"
                updated: "2024-06-02T12:00:00Z"
                isActive: true
        "404":
          $ref: "#/components/responses/NotFoundError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
    delete:
      summary: Delete department
      description: Admin only - Delete a department
      operationId: deleteDepartment
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Department deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Department deleted successfully"
              example:
                message: "Department deleted successfully"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"

  # Employees API
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    UnauthorizedError:
      description: Access token is missing or invalid, or the user does not have access to perform the action
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: "Unauthorized"
    NotFoundError:
      description: Not Found
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: "Not Found"
  schemas:
    Department:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "IT Department"
        description:
          type: string
          example: "Information Technology Department"
        code:
          type: string
          example: "IT"
        location:
          type: string
          example: "Floor 3"
        accountId:
          type: integer
          example: 1
        created:
          type: string
          format: date-time
        updated:
          type: string
          format: date-time
        isActive:
          type: boolean
    CreateDepartmentRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          example: "IT Department"
        description:
          type: string
          example: "Information Technology Department"
        code:
          type: string
          example: "IT"
        location:
          type: string
          example: "Floor 3"
        accountId:
          type: integer
          example: 1
    UpdateDepartmentRequest:
      type: object
      properties:
        name:
          type: string
          example: "IT Department"
        description:
          type: string
          example: "Information Technology Department"
        code:
          type: string
          example: "IT"
        location:
          type: string
          example: "Floor 3"
        accountId:
          type: integer
          example: 1
...